requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anthropic==0.3.11
anyio==3.7.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.2
blinker==1.9.0
Brotli==1.1.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.18
click==8.1.7
coloredlogs==15.0.1
cssselect2==0.7.0
Deprecated==1.2.15
distro==1.9.0
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.7.1
json5==0.9.14
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kubernetes==31.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
narwhals==1.14.0
numpy==2.1.3
oauthlib==3.2.2
onnxruntime==1.20.0
openai==0.28.1
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
orjson==3.10.11
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfkit==1.0.0
pillow==10.4.0
posthog==3.7.0
propcache==0.2.0
protobuf==5.28.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
PyMuPDF==1.24.13
PyPDF2==3.0.1
pyphen==0.15.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
reportlab==4.2.5
requests==2.31.0
requests-oauthlib==2.0.0
# Editable install with no version control (resume_matcher==0.1)
-e /Users/mac/PycharmProjects/re-op-new
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
termcolor==2.3.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.1.0
tornado==6.4.1
tqdm==4.66.1
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.25.11
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
weasyprint==63.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wrapt==1.16.0
yarl==1.17.2
zipp==3.21.0
zopfli==0.2.3.post1
